Basic refactor for names to folders and .dart files when use the new route approach with modular
