+++ lib/app/routing/app_route_module.dart
@@ -1,10 +1,10 @@
 import 'package:flutter_modular/flutter_modular.dart';
-import 'package:portugal_guide/app/rouute_main_stuff/custom_route_manager.dart';
+import 'package:portugal_guide/app/routing/app_route_manager.dart';
 import 'package:portugal_guide/features/core_auth/screens/core_auth_login_screen.dart';
 import 'package:portugal_guide/features/core_auth/screens/core_auth_register_screen.dart';
 import 'app_routes.dart';
 
-class AppModule extends Module {
+class AppRouteModule extends Module {
   @override
   void routes(RouteManager r) {
     final routes = {
@@ -21,7 +21,7 @@ class AppModule extends Module {
     CustomRouteManager.setupRoutes(
       routes: routes,
       routeManager: r,
-      guards: AppRoutes.guardedRoutes,
+      guards: AppRoutes.basicAuthGuardGroupsTest,
     );
   }
 }
