+++ lib/app/routing_guards/owner_content_guard.dart
@@ -1,0 +1,3 @@
+class OwnerContentGuard {
+  //TODO --> FINISH THIS IMPLEMENTATION
+}
